moved functions related to scan management into their own module
