+++ src/main.rs
@@ -3,7 +3,8 @@ use feroxbuster::{
     banner,
     config::{CONFIGURATION, PROGRESS_BAR, PROGRESS_PRINTER},
     heuristics, logger, reporter,
-    scanner::{self, scan_url, PAUSE_SCAN},
+    scanner::{self, scan_url},
+    scan_manager::PAUSE_SCAN,
     utils::{ferox_print, get_current_depth, module_colorizer, status_colorizer},
     FeroxError, FeroxResponse, FeroxResult, SLEEP_DURATION, VERSION,
 };
